refactor collection conditions
