+++ src/packages/block/block/types.ts
@@ -1,5 +1,6 @@
 import type { UmbContentValueModel } from '@umbraco-cms/backoffice/content';
 
+export type * from './conditions/types.js';
 export type * from './clipboard/types.js';
 
 export interface UmbBlockLayoutBaseModel {
