+++ src/packages/core/collection/conditions/collection-alias.condition.ts
@@ -1,5 +1,5 @@
-import { UMB_COLLECTION_CONTEXT } from './default/index.js';
-import type { CollectionAliasConditionConfig } from './collection-alias.manifest.js';
+import { UMB_COLLECTION_CONTEXT } from '../default/index.js';
+import type { CollectionAliasConditionConfig } from './types.js';
 import type { UmbConditionControllerArguments, UmbExtensionCondition } from '@umbraco-cms/backoffice/extension-api';
 import { UmbConditionBase } from '@umbraco-cms/backoffice/extension-registry';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
