+++ src/packages/core/collection/conditions/collection-bulk-action-permission.condition.ts
@@ -1,5 +1,5 @@
-import { UMB_COLLECTION_CONTEXT } from './default/index.js';
-import type { CollectionBulkActionPermissionConditionConfig } from './collection-bulk-action-permission.manifest.js';
+import { UMB_COLLECTION_CONTEXT } from '../default/index.js';
+import type { CollectionBulkActionPermissionConditionConfig } from './types.js';
 import type { UmbConditionControllerArguments, UmbExtensionCondition } from '@umbraco-cms/backoffice/extension-api';
 import { UmbConditionBase } from '@umbraco-cms/backoffice/extension-registry';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
