+++ src/packages/core/collection/conditions/constants.ts
@@ -1,0 +1,3 @@
+export const UMB_COLLECTION_ALIAS_CONDITION = 'Umb.Condition.CollectionAlias';
+
+export const UMB_COLLECTION_BULK_ACTION_PERMISSION_CONDITION = 'Umb.Condition.CollectionBulkActionPermission';
