+++ src/packages/core/collection/index.ts
@@ -5,15 +5,13 @@ import './components/index.js';
 export * from './default/collection-default.element.js';
 export * from './collection.element.js';
 export * from './components/index.js';
+export * from './conditions/index.js';
 export * from './collection-item-picker-modal/index.js';
 
 export * from './default/collection-default.context.js';
 export * from './default/collection-default.context-token.js';
 export * from './collection-filter-model.interface.js';
 export type * from './types.js';
-
-export { UMB_COLLECTION_ALIAS_CONDITION } from './collection-alias.manifest.js';
-export { UMB_COLLECTION_BULK_ACTION_PERMISSION_CONDITION } from './collection-bulk-action-permission.manifest.js';
 
 export { UmbCollectionActionElement, UmbCollectionActionBase } from './action/index.js';
 export type { UmbCollectionDataSource, UmbCollectionRepository } from './repository/index.js';
