+++ src/packages/core/collection/types.ts
@@ -3,6 +3,7 @@ import type { Observable } from '@umbraco-cms/backoffice/external/rxjs';
 import type { UmbPaginationManager } from '@umbraco-cms/backoffice/utils';
 
 export type * from './extensions/index.js';
+export type * from './conditions/types.js';
 
 export interface UmbCollectionBulkActionPermissions {
 	allowBulkCopy: boolean;
